include: Move file replacement translation
Include line replacement application now lives in
commands.selection.include_line_replacement, while include.py still owns
file-view replacement ID translation.

The translation path uses the same replacement expansion and annotated
source identities as replacement application, so keeping it in the
entrypoint splits one include --line --as support flow across two modules.

This commit continues the line-replacement extraction by moving file-view
replacement ID translation and its identity helper into
commands.selection.include_line_replacement.

The next commit will guard the include line-replacement helper boundary in
the architecture suite.
